+++ theme/js/components/dataset/resource/edit-button.vue
@@ -2,14 +2,14 @@
   <a
     :href="adminUrl"
     :title="$t('Edit resource')"
-    class="fr-btn fr-btn--secondary fr-btn--secondary-grey-500 fr-btn--sm fr-icon-svg"
-    v-html="edit"
-    >
+    class="fr-btn fr-btn--secondary fr-btn--secondary-grey-500 fr-btn--sm fr-icon-svg fr-icon--sm"
+  >
+    <img :src="edit" alt="" />
   </a>
 </template>
 
 <script>
-import edit from "bundle-text:svg/edit.svg";
+import edit from "svg/edit.svg";
 import config from "../../../config";
 
 export default {