+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## Current (in progress)
 
-- Nothing yet
+> **Note** <br>
+> This changes the build system from Parcel to Vite.
+> This also adds a way for udata plugin to register their own vue components
+> and to display them in places chosen by the current theme
+
+- Add resource exploration preview [#169](https://github.com/etalab/udata-front/pull/169)
 
 ## 3.0.1 (2022-12-15)